updates on new db tables - API is currently not working as it needs to be re-tuned for new structure

+++ mb40mktg/database/migrations/2018_08_14_113819_add_api_token_to_users_table.php
@@ -25,8 +25,10 @@ class AddApiTokenToUsersTable extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn(['api_token']);
-        });
+        if (Schema::hasTable('users')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn(['api_token']);
+            });
+        }
     }
 }
